background-removal-app/requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
distlib==0.3.8
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.1
pyfiglet==1.0.2
PyMatting==1.1.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
referencing==0.32.0
rembg==2.0.53
requests==2.31.0
rpds-py==0.16.2
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
spotipy==2.23.0
sympy==1.12
tifffile==2023.12.9
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
